py/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
joblib==1.3.2
loguru==0.7.2
numpy==1.26.4
pandas==2.2.1
protobuf==4.25.3
python-dateutil==2.8.2
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
threadpoolctl==3.3.0
tzdata==2024.1
